+++ server/src/core/api-rest/api-rest-query-builder/factories/factories.ts
@@ -7,10 +7,10 @@ import { DeleteVariablesFactory } from 'src/core/api-rest/api-rest-query-builder
 import { CreateVariablesFactory } from 'src/core/api-rest/api-rest-query-builder/factories/create-variables.factory';
 import { UpdateVariablesFactory } from 'src/core/api-rest/api-rest-query-builder/factories/update-variables.factory';
 import { GetVariablesFactory } from 'src/core/api-rest/api-rest-query-builder/factories/get-variables.factory';
-import { LastCursorParserFactory } from 'src/core/api-rest/api-rest-query-builder/factories/last-cursor-parser.factory';
-import { LimitParserFactory } from 'src/core/api-rest/api-rest-query-builder/factories/limit-parser.factory';
-import { OrderByParserFactory } from 'src/core/api-rest/api-rest-query-builder/factories/order-by-parser.factory';
-import { FilterParserFactory } from 'src/core/api-rest/api-rest-query-builder/factories/filter-parser.factory';
+import { LastCursorParserFactory } from 'src/core/api-rest/api-rest-query-builder/factories/parsers/last-cursor-parser.factory';
+import { LimitParserFactory } from 'src/core/api-rest/api-rest-query-builder/factories/parsers/limit-parser.factory';
+import { OrderByParserFactory } from 'src/core/api-rest/api-rest-query-builder/factories/parsers/order-by-parser.factory';
+import { FilterParserFactory } from 'src/core/api-rest/api-rest-query-builder/factories/parsers/filter-parser.factory';
 
 export const apiRestQueryBuilderFactories = [
   DeleteQueryFactory,
